Fix synchronization.
synchronize is renamed to sync. join and sync aren't as common as
originally thought, so they are rewritten with that in mind.

New method added called wait; similar to a join + sync. It will bring
all threads to a synchronized position, but doesn't delete the workers.
It will also wait in the calling thread until the synchronization finishes.

Fix condition variable predicates; they need to be negated.

Rewrite how to add jobs to the work queue.