Fix issues with how workers are created and managed.
Make WorkerThread and ThreadPool sharable. This allows references
to be replaced with shared pointers to ensure that these objects
cannot go out of scope.

Change CreateWorkerKey to StartWorkerKey to represent how workers
can be created, but not started without a thread pool owning it.
This change was made because it is easier to restrict the starting
of a worker than it is the creation of a worker. Creating workers
has the added complexity of additional parameters being provided for
user defined worker thread types which would mean giving the user
access to the CreateWorkerKey in the factory; giving up complete
control of the key.