Add sample thread pools.
PausableThreadPool adds pause & resume methods.

StaticThreadPool doesn't allow workers to be added or deleted except
for invalidating the pool by a call to join.

Add Gate which is similar to Barrier, but blocks everything until
opened separately.

Replace an atomic variable in Barrier with a regular variable as
all operations with it are protected by a mutex.